[PM-21676] Relocate local manager providers
Move the `CompositionLocal` providers for various managers (PermissionsManager, IntentManager, ExitManager, BiometricsManager) from `AuthenticatorTheme.kt` to a new dedicated composable `LocalManagerProvider.kt`.

This change improves code organization by centralizing the provision of these managers. Screen composables have been updated to import these local providers from the new `composition` package.
