+++ authenticator/src/main/kotlin/com/bitwarden/authenticator/ui/platform/theme/AuthenticatorTheme.kt
@@ -20,14 +20,6 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
 import androidx.core.view.WindowCompat
 import com.bitwarden.authenticator.R
-import com.bitwarden.authenticator.ui.platform.manager.biometrics.BiometricsManager
-import com.bitwarden.authenticator.ui.platform.manager.biometrics.BiometricsManagerImpl
-import com.bitwarden.authenticator.ui.platform.manager.exit.ExitManager
-import com.bitwarden.authenticator.ui.platform.manager.exit.ExitManagerImpl
-import com.bitwarden.authenticator.ui.platform.manager.intent.IntentManager
-import com.bitwarden.authenticator.ui.platform.manager.intent.IntentManagerImpl
-import com.bitwarden.authenticator.ui.platform.manager.permissions.PermissionsManager
-import com.bitwarden.authenticator.ui.platform.manager.permissions.PermissionsManagerImpl
 import com.bitwarden.ui.platform.feature.settings.appearance.model.AppTheme
 
 /**
@@ -47,7 +39,6 @@ fun AuthenticatorTheme(
 
     // Get the current scheme
     val context = LocalContext.current
-    val activity = context as Activity
     val colorScheme = when {
         dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
             if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
@@ -79,10 +70,6 @@ fun AuthenticatorTheme(
     CompositionLocalProvider(
         LocalNonMaterialColors provides nonMaterialColors,
         LocalNonMaterialTypography provides nonMaterialTypography,
-        LocalPermissionsManager provides PermissionsManagerImpl(activity),
-        LocalIntentManager provides IntentManagerImpl(context),
-        LocalExitManager provides ExitManagerImpl(activity),
-        LocalBiometricsManager provides BiometricsManagerImpl(activity),
     ) {
         // Set overall theme based on color scheme and typography settings
         MaterialTheme(
@@ -170,34 +157,6 @@ private fun lightColorScheme(context: Context): ColorScheme =
 @ColorRes
 private fun Int.toColor(context: Context): Color =
     Color(context.getColor(this))
-
-/**
- * Provides access to the biometrics manager throughout the app.
- */
-val LocalBiometricsManager: ProvidableCompositionLocal<BiometricsManager> = compositionLocalOf {
-    error("CompositionLocal BiometricsManager not present")
-}
-
-/**
- * Provides access to the exit manager throughout the app.
- */
-val LocalExitManager: ProvidableCompositionLocal<ExitManager> = compositionLocalOf {
-    error("CompositionLocal ExitManager not present")
-}
-
-/**
- * Provides access to the intent manager throughout the app.
- */
-val LocalIntentManager: ProvidableCompositionLocal<IntentManager> = compositionLocalOf {
-    error("CompositionLocal LocalIntentManager not present")
-}
-
-/**
- * Provides access to the permission manager throughout the app.
- */
-val LocalPermissionsManager: ProvidableCompositionLocal<PermissionsManager> = compositionLocalOf {
-    error("CompositionLocal LocalPermissionsManager not present")
-}
 
 /**
  * Provides access to non material theme typography throughout the app.
